Move specific u64_backend values to it's constants spec file
